+++ CHANGELOG.md
@@ -10,7 +10,11 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Updated mongo-lib to v4.5.2
 
 ### Changes
+- dbcc, defrag: Replaced get_db_tbl call with mongo_libs.get_db_tbl call.
 - Documentation changes.
+
+### Removed
+- get_db_tbl function
 
 
 ## [4.1.2] - 2025-03-18